+++ SceneManagement/AdditiveSceneManager.cs
@@ -6,8 +6,10 @@ using System.IO;
 using System.Linq;
 using UnityEngine;
 using UnityEngine.SceneManagement;
+
 #if UNITY_EDITOR
 using UnityEditor;
+using UnityEditor.SceneManagement;
 #endif
 
 namespace SceneManagement
@@ -15,7 +17,7 @@ namespace SceneManagement
     [DefaultExecutionOrder(-10000)]
     public class AdditiveSceneManager : MonoBehaviour
     {
-        [field: SerializeField, ValueDropdown("GetAllScenes")] public List<string> SceneList { get; private set; }
+        [field: SerializeField, ValueDropdown("GetAllScenes"), InlineButton("AddNewScene", "New")] public List<string> SceneList { get; private set; }
 
         private void Awake()
         {
@@ -70,6 +72,39 @@ namespace SceneManagement
             }
 
             return false;
+        }
+
+        private void AddNewScene()
+        {
+#if UNITY_EDITOR
+            Scene activeScene = SceneManager.GetActiveScene();
+            string path = activeScene.path.Replace(activeScene.name + ".unity", "");
+            Scene newScene = EditorSceneManager.NewScene(NewSceneSetup.EmptyScene, NewSceneMode.Additive);
+
+            int index = GetLoadedSceneIndex(newScene);
+            if (index == -1)
+            {
+                Debug.LogError("Scene failed to load correctly");
+                return;
+            }
+
+            newScene.name = $"{activeScene.name}-{index}";
+            EditorSceneManager.SaveScene(newScene, path + "/" + newScene.name + ".unity");
+            EditorSceneManager.SetActiveScene(activeScene);
+            SceneList.Add(newScene.path);
+
+            Debug.Log($"New scene added: {newScene.name}");
+#endif
+        }
+
+        private int GetLoadedSceneIndex(Scene scene)
+        {
+            for(int i = 0; i < SceneManager.loadedSceneCount; i++)
+            {
+                if (SceneManager.GetSceneAt(i) == scene) return i;
+            }
+
+            return -1;
         }
 
 #if UNITY_EDITOR